Changed authorization of editing and deletion of uploaded image by author only

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -129,39 +129,34 @@ public class ImageController {
         return "redirect:/images";
     }
 
-    //This controller method is called when the request pattern is of type 'editImage'
-    //This method fetches the image with the corresponding id from the database and adds it to the model with the key as 'image'
-    //The method then returns 'images/edit.html' file wherein you fill all the updated details of the image
 
-    //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
-    //This string is then displayed by 'edit.html' file as previous tags of an image
-    @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
-        Image image = imageService.getImage(imageId);
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
-    }
-
-    //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
-    //The method receives the imageFile, imageId, updated image, along with the Http Session
-    //The method adds the new imageFile to the updated image if user updates the imageFile and adds the previous imageFile to the new updated image if user does not choose to update the imageFile
-    //Set an id of the new updated image
-    //Set the user using Http Session
-    //Set the date on which the image is posted
-    //Call the updateImage() method in the business logic to update the image
-    //Direct to the same page showing the details of that particular updated image
-
-    //The method also receives tags parameter which is a string of all the tags separated by a comma using the annotation @RequestParam
-    //The method converts the string to a list of all the tags using findOrCreateTags() method and sets the tags attribute of an image as a list of all the tags
+    /**
+     * This request handler method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
+     * The method receives the imageFile, imageId, tags, updated image, along with the Http Session
+     * The method first converts the string of all tags separated by a comma to a list of all the tags using findOrCreatetags() method
+     * Set the tags attribute of the image as the list of tags returned by the findOrCreateTags() method
+     * The method adds the new imageFile to the updated image if user updates the imageFile and adds the previous imageFile to the new updated image if user does not choose to update the imageFile
+     * Set an id of the new updated image
+     * Set the user using Http Session
+     * Set the date on which the image is posted
+     * Call the updateImage() method in the business logic to update the image
+     * Direct to the same page showing the details of that particular updated image
+     *
+     * @param file         - This request parameter contains the image file
+     * @param imageId      - This request parameter contains the id of the edited image
+     * @param tags         - This request parameter contains the string of all tags separated by a comma
+     * @param updatedImage - An object of type Image containing the updated attributes of the image. Note that the image attributes which are not updated are become null now
+     * @param session      - Http session containing the details of the logged in user
+     * @return - Returns the same page showing the details of that particular updated image
+     * @throws IOException
+     */
     @RequestMapping(value = "/editImage", method = RequestMethod.PUT)
     public String editImageSubmit(@RequestParam("file") MultipartFile file, @RequestParam("imageId") Integer imageId, @RequestParam("tags") String tags, Image updatedImage, HttpSession session) throws IOException {
 
         Image image = imageService.getImage(imageId);
-        String updatedImageData = convertUploadedFileToBase64(file);
         List<Tag> imageTags = findOrCreateTags(tags);
+        String updatedImageData = convertUploadedFileToBase64(file);
 
         if (updatedImageData.isEmpty())
             updatedImage.setImageFile(image.getImageFile());
@@ -176,28 +171,61 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + updatedImage.getId() + "/" + updatedImage.getTitle();
     }
 
 
-    //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
-    //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
-    //Looks for a controller method with request mapping of type '/images'
+    /**
+     * This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
+     * The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
+     * Looks for a controller method with request mapping of type '/images'
+     * You also need to check whether the user who is deleting the image is the owner of the image or not?
+     * How to do that?
+     * Get the current logged in user from the session and compare that user with the owner of the image
+     * If the current logged in user matches with the owner of the image, user can delete the image and the method redirects to the handling method with request mapping of type '/images'
+     * Else you need to print the error message
+     * And add the deleteError, image, tags and comments in the Model type object and return the 'images/image.html' file again showing the details of the image
+     *
+     * @param imageId - This request parameter contains the id of the image to be deleted
+     * @param model   - model is an object of Type Model, a class provided by the Spring. You can add the attributes in this Model type object and then access these attributes in the HTML files
+     * @param session - Http session containing the details of the logged in user
+     * @return - This method redirects to the handling method with request mapping of type '/images' if the current user is the owner of the image. Else, the method redirects to 'images/image.html' file
+     */
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        User user = (User) (session.getAttribute("loggeduser"));
+        Image image = imageService.getImage(imageId);
+        if ((user.getId()).equals(image.getUser().getId())) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else {
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError", error);
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
+
     }
 
 
-    //This method converts the image to Base64 format
+    /**
+     * This method converts the image to Base64 format
+     *
+     * @param file - The image file
+     * @return - This method returns the Base64 encoded image as a string
+     * @throws IOException
+     */
     private String convertUploadedFileToBase64(MultipartFile file) throws IOException {
         return Base64.getEncoder().encodeToString(file.getBytes());
     }
 
-    //findOrCreateTags() method has been implemented, which returns the list of tags after converting the ‘tags’ string to a list of all the tags and also stores the tags in the database if they do not exist in the database. Observe the method and complete the code where required for this method.
-    //Try to get the tag from the database using getTagByName() method. If tag is returned, you need not to store that tag in the database, and if null is returned, you need to first store that tag in the database and then the tag is added to a list
-    //After adding all tags to a list, the list is returned
+    /**
+     * The method returns the list of all the tags and persists the non existing tags in the database
+     *
+     * @param tagNames - A string containing all the tags of the image separated by a comma
+     * @return - The list of all the tags
+     */
     private List<Tag> findOrCreateTags(String tagNames) {
         StringTokenizer st = new StringTokenizer(tagNames, ",");
         List<Tag> tags = new ArrayList<Tag>();
@@ -215,10 +243,14 @@ public class ImageController {
         return tags;
     }
 
-
-    //The method receives the list of all tags
-    //Converts the list of all tags to a single string containing all the tags separated by a comma
-    //Returns the string
+    /**
+     * The method receives the list of all tags
+     * Converts the list of all tags to a single string containing all the tags separated by a comma
+     * Returns the string
+     *
+     * @param tags - The list of all the tags
+     * @return - The method returns the string containing all the tags separated by a comma
+     */
     private String convertTagsToString(List<Tag> tags) {
         StringBuilder tagString = new StringBuilder();
